+++ include/helib/Ctxt.h
@@ -54,10 +54,12 @@
  * ciphertext wrt (1,s(X^2)) to another wrt (1,s(X^2),s^2(X^2)), but not
  * to another ciphertext wrt (1,s).
  **/
-#include <cfloat> // DBL_MAX
+
 #include <helib/DoubleCRT.h>
 #include <helib/EncodedPtxt.h>
 #include <helib/apiAttributes.h>
+
+#include <cfloat> // DBL_MAX
 
 namespace helib {
 struct CKKS;
@@ -568,22 +570,12 @@ public:
   // Multiply by another ciphertext
   void multLowLvl(const Ctxt& other, bool destructive = false);
 
-#if 0
-  [[deprecated]]
-  Ctxt& operator*=(const Ctxt& other)
-  {
-    multLowLvl(other);
-    return *this;
-  }
-#else
-  // we now do the high-level mul
+  // This is a high-level mul with relinearization
   Ctxt& operator*=(const Ctxt& other)
   {
     multiplyBy(other);
     return *this;
   }
-
-#endif
 
   void automorph(long k); // Apply automorphism F(X) -> F(X^k) (gcd(k,m)=1)
   Ctxt& operator>>=(long k)
@@ -612,8 +604,8 @@ public:
    * Please use `Ctxt::operator*=(const EncodedPtxt& ptxt)` instead.
    **/
   [[deprecated(
-      "Please use Ctxt::operator*=(const EncodedPtxt& ptxt) instead.")]]
-  Ctxt& operator*=(const NTL::ZZX& poly);
+      "Please use Ctxt::operator*=(const EncodedPtxt& ptxt) instead.")]] Ctxt&
+  operator*=(const NTL::ZZX& poly);
 
   //! Add a constant polynomial.
   //! If provided, size should be a high-probability bound
@@ -656,18 +648,19 @@ public:
    * @deprecated This function is deprecated in favor of a newer API.
    * Please use `Ctxt::operator+=(double ptxt)` instead.
    **/
-  [[deprecated("Please use Ctxt::operator+=(double ptxt) instead.")]]
-  void addConstantCKKS(std::pair</*numerator=*/long, /*denominator=*/long>);
+  [[deprecated("Please use Ctxt::operator+=(double ptxt) instead.")]] void
+      addConstantCKKS(std::pair</*numerator=*/long, /*denominator=*/long>);
   /**
    * @deprecated This function is deprecated in favor of a newer API.
    * Please use `Ctxt::operator+=(double ptxt)` instead.
    **/
-  [[deprecated("Please use Ctxt::operator+=(double ptxt) instead.")]]
-  void addConstantCKKS(double x)
+  [[deprecated("Please use Ctxt::operator+=(double ptxt) instead.")]] void
+  addConstantCKKS(double x)
   { // FIXME: not enough precision when x is large
     // This function is deprecated.
-//    addConstantCKKS(
-//        rationalApprox(x, /*denomBound=*/1L << getContext().getAlMod().getR()));
+    //    addConstantCKKS(
+    //        rationalApprox(x, /*denomBound=*/1L <<
+    //        getContext().getAlMod().getR()));
     auto p =
         rationalApprox(x, /*denomBound=*/1L << getContext().getAlMod().getR());
     *this += double(p.first) / p.second;
@@ -687,17 +680,18 @@ public:
    * Please use `Ctxt::operator+=(const EncodedPtxt& ptxt)` instead.
    **/
   [[deprecated(
-      "Please use Ctxt::operator+=(const EncodedPtxt& ptxt) instead.")]]
-  void addConstantCKKS(const NTL::ZZX& poly,
-                       NTL::xdouble size = NTL::xdouble(-1.0),
-                       NTL::xdouble factor = NTL::xdouble(-1.0));
+      "Please use Ctxt::operator+=(const EncodedPtxt& ptxt) instead.")]] void
+  addConstantCKKS(const NTL::ZZX& poly,
+                  NTL::xdouble size = NTL::xdouble(-1.0),
+                  NTL::xdouble factor = NTL::xdouble(-1.0));
 
   /**
    * @deprecated This function is deprecated in favor of a newer API.
    * Please use `Ctxt::operator+=(const PtxtArray& ptxt)` instead.
    **/
-  [[deprecated("Please use Ctxt::operator+=(const PtxtArray& ptxt) instead.")]]
-  void addConstantCKKS(const std::vector<std::complex<double>>& ptxt);
+  [[deprecated(
+      "Please use Ctxt::operator+=(const PtxtArray& ptxt) instead.")]] void
+  addConstantCKKS(const std::vector<std::complex<double>>& ptxt);
 
   /**
    * @brief Add a `CKKS` plaintext to this `Ctxt`.
@@ -706,14 +700,15 @@ public:
    * Please use `Ctxt::operator+=(const PtxtArray& ptxt)` instead.
    **/
   [[deprecated(
-      "Please use Ctxt::operator+=(const Ptxt<Scheme>& ptxt) instead.")]]
-  void addConstantCKKS(const Ptxt<CKKS>& ptxt);
+      "Please use Ctxt::operator+=(const Ptxt<Scheme>& ptxt) instead.")]] void
+  addConstantCKKS(const Ptxt<CKKS>& ptxt);
   /**
    * @deprecated This function is deprecated in favor of a newer API.
    * Please use `Ctxt::operator+=(const NTL::ZZ& ptxt)` instead.
    **/
-  [[deprecated("Please use Ctxt::operator+=(const NTL::ZZ& ptxt) instead.")]]
-  void addConstantCKKS(const NTL::ZZ& c);
+  [[deprecated(
+      "Please use Ctxt::operator+=(const NTL::ZZ& ptxt) instead.")]] void
+  addConstantCKKS(const NTL::ZZ& c);
 
   //! Multiply-by-constant.
   //! If the size is not given, for the DCRT variant, we use
@@ -1101,8 +1096,8 @@ public:
    * @deprecated This function is deprecated in favor of a newer API.
    * Please use `Ctxt::operator*=(double ptxt)` instead.
    **/
-  [[deprecated("Please use Ctxt::operator*=(double ptxt) instead.")]]
-  void multByConstantCKKS(double x)
+  [[deprecated("Please use Ctxt::operator*=(double ptxt) instead.")]] void
+  multByConstantCKKS(double x)
   {
     if (this->isEmpty())
       return;
@@ -1126,8 +1121,8 @@ public:
    * @deprecated This function is deprecated in favor of a newer API.
    * Please use `Ctxt::operator*=(double ptxt)` instead.
    **/
-  [[deprecated("Please use Ctxt::operator*=(double ptxt) instead.")]]
-  void multByConstantCKKS(std::pair<long, long> num) // rational number
+  [[deprecated("Please use Ctxt::operator*=(double ptxt) instead.")]] void
+  multByConstantCKKS(std::pair<long, long> num) // rational number
   {
     // This function is deprecated.
     // multByConstantCKKS(double(num.first) / num.second);
@@ -1149,11 +1144,11 @@ public:
    * Please use `Ctxt::operator*=(const EncodedPtxt& ptxt)` instead.
    **/
   [[deprecated(
-      "Please use Ctxt::operator*=(const EncodedPtxt& ptxt) instead.")]]
-  void multByConstantCKKS(const NTL::ZZX& poly,
-                          NTL::xdouble size = NTL::xdouble(-1.0),
-                          NTL::xdouble factor = NTL::xdouble(-1.0),
-                          double roundingErr = -1.0)
+      "Please use Ctxt::operator*=(const EncodedPtxt& ptxt) instead.")]] void
+  multByConstantCKKS(const NTL::ZZX& poly,
+                     NTL::xdouble size = NTL::xdouble(-1.0),
+                     NTL::xdouble factor = NTL::xdouble(-1.0),
+                     double roundingErr = -1.0)
   {
     DoubleCRT dcrt(poly, context, primeSet);
     multByConstantCKKS(dcrt, size, factor, roundingErr);
@@ -1168,15 +1163,15 @@ public:
    * Please use `Ctxt::operator*=(const PtxtArray& ptxt)` instead.
    **/
   [[deprecated(
-      "Please use Ctxt::operator*=(const Ptxt<Scheme>& ptxt) instead.")]]
-  void multByConstantCKKS(const Ptxt<CKKS>& ptxt);
+      "Please use Ctxt::operator*=(const Ptxt<Scheme>& ptxt) instead.")]] void
+  multByConstantCKKS(const Ptxt<CKKS>& ptxt);
   /**
    * @deprecated This function is deprecated in favor of a newer API.
    * Please use `Ctxt::operator*=(const PtxtArray& ptxt)` instead.
    **/
   [[deprecated(
-      "Please use Ctxt::operator*=(const PtxtArray& ptxt) instead.")]]
-  void multByConstantCKKS(const std::vector<std::complex<double>>& ptxt);
+      "Please use Ctxt::operator*=(const PtxtArray& ptxt) instead.")]] void
+  multByConstantCKKS(const std::vector<std::complex<double>>& ptxt);
 
   //! Convenience method: XOR and nXOR with arbitrary plaintext space:
   //! a xor b = a+b-2ab = a + (1-2a)*b,
@@ -1389,25 +1384,14 @@ public:
   bool isCKKS() const { return getContext().isCKKS(); }
 
   // Return r such that p^r = ptxtSpace
-  long effectiveR() const
-  {
-    long p = context.getP();
-    for (long r = 1, p2r = p; r < NTL_SP_NBITS; r++, p2r *= p) {
-      if (p2r == ptxtSpace)
-        return r;
-      if (p2r > ptxtSpace)
-        throw RuntimeError("ctxt.ptxtSpace is not of the form p^r");
-    }
-    throw RuntimeError("ctxt.ptxtSpace is not of the form p^r");
-    return 0; // just to keep the compiler happy
-  }
+  long effectiveR() const;
 
   /**
    * @brief Returns log(noiseBound) - log(q)
    * @deprecated This is deprecated. Please use `Ctxt::capacity()` instead.
    **/
-  [[deprecated("Please use Ctxt::capacity() instead.")]]
-  double log_of_ratio() const
+  [[deprecated("Please use Ctxt::capacity() instead.")]] double log_of_ratio()
+      const
   {
     double logNoise =
         (getNoiseBound() <= 0.0) ? -DBL_MAX : log(getNoiseBound());