+++ misc/format.sh
@@ -16,13 +16,14 @@ FORMAT_PRG="${1:-"clang-format"}"
 FORMAT_CMD="${FORMAT_PRG} -style=file -i"
 
 # Check we can find clang-format
-if [ -z $(which $FORMAT_PRG) ]; then
+if [ -z "$(which $FORMAT_PRG)" ]; then
   echo "$FORMAT_PRG not found."
   exit 1
 fi
 
 # Check format program is correct major version
-if [ ! $($FORMAT_PRG --version | cut -d' ' -f3 | cut -d'.' -f1) -ge 9 ]; then
+# NOTE Some OSes print extra info, but the version number appears at the end.
+if [ ! $($FORMAT_PRG --version | awk '{print $NF}' | cut -d'.' -f1) -ge 9 ]; then
   >&2 echo "Clang-format version below 9. Require 9+."
   exit 2
 fi
@@ -43,8 +44,14 @@ function echo_header {
 
 echo "Using '${FORMAT_CMD}' to perform formatting."
 
+arg="-regextype posix-egrep"
+# find command differs between Mac and Linux
+if [ "$(uname)" == "Darwin" ]; then
+  arg="-E"
+fi
+
 previous_dir=""
-for file in $(find -E . -type f -regex '.*\.(c|h|cpp|hpp)' \
+for file in $(find . $arg -type f -regex '.*\.(c|h|cpp|hpp)' \
               ! -path '*/misc/*' \
               ! -path '*/build/*' \
               ! -path '*/dependencies/*' \