+++ src/DoubleCRT.cpp
@@ -9,6 +9,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License. See accompanying LICENSE file.
  */
+
+/* Intel HEXL integration.
+ * Copyright (C) 2021 Intel Corporation
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 /* DoubleCRT.cpp - This class holds an integer polynomial in double-CRT form
  *
  * Double-CRT form is a matrix of L rows and phi(m) columns. The i'th row
@@ -23,6 +37,7 @@
 
 #include "binio.h"
 #include "io.h"
+#include "intelExt.h"
 
 #include <helib/timing.h>
 #include <helib/sample.h>
@@ -117,8 +132,85 @@ void DoubleCRT::verify()
   }
 }
 
-// Arithmetic operations. Only the "destructive" versions are used,
-// i.e., a += b is implemented but not a + b.
+#ifdef USE_INTEL_HEXL
+struct AddFun
+{
+  void apply(long* result,
+             const long* a,
+             const long* b,
+             long size,
+             long modulus) const
+  {
+    intel::EltwiseAddMod(result, a, b, size, modulus);
+  }
+
+  void apply(long* result,
+             const long* a,
+             long scalar,
+             long size,
+             long modulus) const
+  {
+    intel::EltwiseAddMod(result, a, scalar, size, modulus);
+  }
+};
+
+struct SubFun
+{
+  void apply(long* result,
+             const long* a,
+             const long* b,
+             long size,
+             long modulus) const
+  {
+    intel::EltwiseSubMod(result, a, b, size, modulus);
+  }
+
+  void apply(long* result,
+             const long* a,
+             long scalar,
+             long size,
+             long modulus) const
+  {
+    intel::EltwiseSubMod(result, a, scalar, size, modulus);
+  }
+};
+
+struct MulFun
+{
+  void apply(long* result,
+             const long* a,
+             const long* b,
+             long size,
+             long modulus) const
+  {
+    intel::EltwiseMultMod(result, a, b, size, modulus);
+  }
+
+  void apply(long* result,
+             const long* a,
+             long scalar,
+             long size,
+             long modulus) const
+  {
+    intel::EltwiseMultMod(result, a, scalar, size, modulus);
+  }
+};
+#else
+struct AddFun
+{
+  long apply(long a, long b, long n) const { return NTL::AddMod(a, b, n); }
+};
+
+struct SubFun
+{
+  long apply(long a, long b, long n) const { return NTL::SubMod(a, b, n); }
+};
+
+struct MulFun
+{
+  long apply(long a, long b, long n) const { return NTL::MulMod(a, b, n); }
+};
+#endif
 
 // Generic operation, Fnc is AddMod, SubMod, or MulMod (from NTL's ZZ module)
 template <typename Fun>
@@ -170,8 +262,12 @@ DoubleCRT& DoubleCRT::Op(const DoubleCRT& other, Fun fun, bool matchIndexSets)
     NTL::vec_long& row = map[i];
     const NTL::vec_long& other_row = (*other_map)[i];
 
+#ifdef USE_INTEL_HEXL
+    fun.apply(row.elts(), row.elts(), other_row.elts(), phim, pi);
+#else
     for (long j : range(phim))
       row[j] = fun.apply(row[j], other_row[j], pi);
+#endif
   }
   return *this;
 }
@@ -226,29 +322,19 @@ DoubleCRT& DoubleCRT::do_mul(const DoubleCRT& other, bool matchIndexSets)
   // add/sub/mul the data, element by element, modulo the respective primes
   for (long i : s) {
     long pi = context.ithPrime(i);
-    NTL::mulmod_t pi_inv = context.ithModulus(i).getQInv();
     NTL::vec_long& row = map[i];
     const NTL::vec_long& other_row = (*other_map)[i];
 
+#ifdef USE_INTEL_HEXL
+    intel::EltwiseMultMod(row.elts(), row.elts(), other_row.elts(), phim, pi);
+#else
+    NTL::mulmod_t pi_inv = context.ithModulus(i).getQInv();
     for (long j : range(phim))
       row[j] = MulMod(row[j], other_row[j], pi, pi_inv);
+#endif // USE_INTEL_HEXL
   }
   return *this;
 }
-
-#if 0
-template
-DoubleCRT& DoubleCRT::Op<DoubleCRT::MulFun>(const DoubleCRT &other, MulFun fun,
-			 bool matchIndexSets);
-#endif
-
-template DoubleCRT& DoubleCRT::Op<DoubleCRT::AddFun>(const DoubleCRT& other,
-                                                     AddFun fun,
-                                                     bool matchIndexSets);
-
-template DoubleCRT& DoubleCRT::Op<DoubleCRT::SubFun>(const DoubleCRT& other,
-                                                     SubFun fun,
-                                                     bool matchIndexSets);
 
 template <typename Fun>
 DoubleCRT& DoubleCRT::Op(const NTL::ZZ& num, Fun fun)
@@ -263,20 +349,16 @@ DoubleCRT& DoubleCRT::Op(const NTL::ZZ& num, Fun fun)
     long pi = context.ithPrime(i);
     long n = rem(num, pi); // n = num % pi
     NTL::vec_long& row = map[i];
+
+#ifdef USE_INTEL_HEXL
+    fun.apply(row.elts(), row.elts(), n, phim, pi);
+#else
     for (long j : range(phim))
       row[j] = fun.apply(row[j], n, pi);
+#endif
   }
   return *this;
 }
-
-template DoubleCRT& DoubleCRT::Op<DoubleCRT::MulFun>(const NTL::ZZ& num,
-                                                     MulFun fun);
-
-template DoubleCRT& DoubleCRT::Op<DoubleCRT::AddFun>(const NTL::ZZ& num,
-                                                     AddFun fun);
-
-template DoubleCRT& DoubleCRT::Op<DoubleCRT::SubFun>(const NTL::ZZ& num,
-                                                     SubFun fun);
 
 DoubleCRT& DoubleCRT::Negate(const DoubleCRT& other)
 {
@@ -313,14 +395,84 @@ DoubleCRT& DoubleCRT::Op(const NTL::ZZX& poly, Fun fun)
   return Op(other, fun);
 }
 
-template DoubleCRT& DoubleCRT::Op<DoubleCRT::MulFun>(const NTL::ZZX& poly,
-                                                     MulFun fun);
+// overloaded ops
+DoubleCRT& DoubleCRT::operator+=(const DoubleCRT& other)
+{
+  return Op(other, AddFun());
+}
 
-template DoubleCRT& DoubleCRT::Op<DoubleCRT::AddFun>(const NTL::ZZX& poly,
-                                                     AddFun fun);
+DoubleCRT& DoubleCRT::operator+=(const NTL::ZZX& poly)
+{
+  return Op(poly, AddFun());
+}
 
-template DoubleCRT& DoubleCRT::Op<DoubleCRT::SubFun>(const NTL::ZZX& poly,
-                                                     SubFun fun);
+DoubleCRT& DoubleCRT::operator+=(const NTL::ZZ& num)
+{
+  return Op(num, AddFun());
+}
+
+DoubleCRT& DoubleCRT::operator+=(long num)
+{
+  return Op(NTL::to_ZZ(num), AddFun());
+}
+
+DoubleCRT& DoubleCRT::operator-=(const DoubleCRT& other)
+{
+  return Op(other, SubFun());
+}
+
+DoubleCRT& DoubleCRT::operator-=(const NTL::ZZX& poly)
+{
+  return Op(poly, SubFun());
+}
+
+DoubleCRT& DoubleCRT::operator-=(const NTL::ZZ& num)
+{
+  return Op(num, SubFun());
+}
+
+DoubleCRT& DoubleCRT::operator-=(long num)
+{
+  return Op(NTL::to_ZZ(num), SubFun());
+}
+
+DoubleCRT& DoubleCRT::operator*=(const DoubleCRT& other)
+{
+  // return Op(other,MulFun());
+  return do_mul(other);
+}
+
+DoubleCRT& DoubleCRT::operator*=(const NTL::ZZX& poly)
+{
+  return Op(poly, MulFun());
+}
+
+DoubleCRT& DoubleCRT::operator*=(const NTL::ZZ& num)
+{
+  return Op(num, MulFun());
+}
+
+DoubleCRT& DoubleCRT::operator*=(long num)
+{
+  return Op(NTL::to_ZZ(num), MulFun());
+}
+
+// Function versions
+void DoubleCRT::Add(const DoubleCRT& other, bool matchIndexSets)
+{
+  Op(other, AddFun(), matchIndexSets);
+}
+
+void DoubleCRT::Sub(const DoubleCRT& other, bool matchIndexSets)
+{
+  Op(other, SubFun(), matchIndexSets);
+}
+
+void DoubleCRT::Mul(const DoubleCRT& other, bool matchIndexSets)
+{
+  // Op(other, MulFun(), matchIndexSets);
+  do_mul(other, matchIndexSets);
+}
 
 // break *this into n digits,according to the primeSets in context.digits
 // returns the sum of the canonical embedding norms of the digits