+++ CHANGES.md
@@ -1,6 +1,19 @@
 Release Changes
 ===============
 
+HElib 2.2.0, September 2021
+=========================
+(tagged as v2.2.0)
+
+* Intel HEXL integration (experimental)
+  * HEXL NTT acceleration for m as a power of 2
+  * HEXL acceleration to fundamental ciphertext operations
+  * Documentation update
+* Update NTL dependency to version 11.5.1
+* Update GMP dependency to version 6.2.1
+* Update to benchmarks
+* Bug fixes
+
 HElib 2.1.0, March 2021
 =========================
 (tagged as v2.1.0)